+++ app/src/behaviors/form.js
@@ -3,9 +3,9 @@ import 'backbone-validation';
 import 'backbone.stickit';
 import Backbone from 'backbone';
 import Marionette from 'backbone.marionette';
-import { props } from 'decorators';
+import { mixin } from 'core-decorators';
 
-@props({
+@mixin({
   defaults: {
     forceUpdate: true,
     tooltip: {