+++ app/src/collections/comments.js
@@ -1,8 +1,8 @@
+import { mixin } from 'core-decorators';
 import Comment from 'models/comment';
 import AppCollection from 'collections/app';
-import { props } from 'decorators';
 
-@props({
+@mixin({
   model: Comment
 })
 export default class Comments extends AppCollection {}