+++ app/src/decorators.js
@@ -1,5 +1,0 @@
-export function props(value) {
-  return function decorator(target) {
-    Object.assign(target.prototype, value);
-  };
-}