+++ app/src/routers/landing.js
@@ -1,8 +1,8 @@
+import { mixin } from 'core-decorators';
 import BaseRouter from 'routers/base';
 import Controller from 'controllers/landing_controller';
-import { props } from 'decorators';
 
-@props({
+@mixin({
   appRoutes: {
     '': 'index'
   },