+++ app/src/views/articles/item.js
@@ -1,12 +1,12 @@
 import Marionette from 'backbone.marionette';
+import { mixin } from 'core-decorators';
 import App from 'application';
 import Session from 'services/session';
 import routes from 'helpers/routes';
 import user from 'helpers/user';
 import template from 'templates/articles/item';
-import { props } from 'decorators';
 
-@props({
+@mixin({
   template,
   tagName: 'article',
   className: 'articles-item',
@@ -21,14 +21,14 @@ import { props } from 'decorators';
   }
 })
 export default class ArticlesItemView extends Marionette.ItemView {
-  articleRemove() {
-    this.model.destroy({ wait: true });
-  }
-
   serializeData() {
     const props = { page: this.options.page };
     Object.assign(props, this.model.toJSON());
 
     return props;
+  }
+
+  articleRemove() {
+    this.model.destroy({ wait: true });
   }
 }