+++ app/src/views/comments/form.js
@@ -1,12 +1,12 @@
 import Marionette from 'backbone.marionette';
+import { mixin } from 'core-decorators';
 import Comment from 'models/comment';
 import FormBehavior from 'behaviors/form';
 import Session from 'services/session';
 import user from 'helpers/user';
 import template from 'templates/comments/form';
-import { props } from 'decorators';
 
-@props({
+@mixin({
   template,
 
   events: {