+++ app/src/views/comments/item.js
@@ -1,9 +1,9 @@
 import Marionette from 'backbone.marionette';
+import { mixin } from 'core-decorators';
 import user from 'helpers/user';
 import template from 'templates/comments/item';
-import { props } from 'decorators';
 
-@props({
+@mixin({
   template,
   className: 'comments-item',
   templateHelpers: { user }