+++ app/src/views/navigation/user.js
@@ -1,10 +1,10 @@
 import Marionette from 'backbone.marionette';
+import { mixin } from 'core-decorators';
 import Session from 'services/session';
 import user from 'helpers/user';
 import template from 'templates/navigation/user';
-import { props } from 'decorators';
 
-@props({
+@mixin({
   template,
   className: 'navbar-user',
   model: Session.currentUser(),